ui: Move some functions to TimelineInnerMetadata
